Split DES table code to separate file. Since similar stop step calculation code was used for both incident and exiting ions, they were unified.
