+++ src/jabs.c
@@ -36,259 +36,6 @@
 #include "stop.h"
 #include "win_compat.h"
 
-extern inline const des *des_table_min_energy_bin(const des_table *dt);
-
-des_table *des_table_init(size_t n) {
-    des_table *dt = malloc(sizeof(des_table));
-    dt->t = NULL;
-    dt->depth_interval_index = NULL;
-    dt->n = 0;
-    dt->n_ranges = 0;
-    dt->depth_increases = TRUE;
-    if(des_table_realloc(dt, n)) {
-        free(dt);
-        return NULL;
-    }
-    return dt;
-}
-int des_table_realloc(des_table *dt, size_t n) {
-    if(n > DES_TABLE_MAX_SIZE) {
-#ifdef DEBUG
-        fprintf(stderr, "DES table requested size %zu larger than allowed %i\n", n, DES_TABLE_MAX_SIZE);
-#endif
-        free(dt->t);
-        dt->t = NULL;
-    } else {
-#ifdef DEBUG
-        fprintf(stderr, "DES table %p realloc to %zu\n", (void *) dt, n);
-#endif
-        dt->t = realloc(dt->t, n * sizeof(des));
-    }
-    if(dt->t) {
-        dt->n = n;
-        return 0;
-    } else {
-        dt->n = 0;
-        return -1;
-    }
-}
-
-void des_table_free(des_table *dt) {
-    if(!dt) {
-        return;
-    }
-    free(dt->t);
-    free(dt->depth_interval_index);
-    free(dt);
-}
-
-size_t des_table_size(const des_table *dt) {
-    if(!dt) {
-        return 0;
-    }
-    return dt->n;
-}
-
-des *des_table_element(const des_table *dt, size_t i) {
-    if(!dt) {
-        return NULL;
-    }
-    if(i > dt->n) {
-        return NULL;
-    }
-    return &(dt->t[i]);
-}
-
-void des_table_rebuild_index(des_table *dt) {
-    if(!dt->n)
-        return;
-    if(dt->depth_interval_index) {
-        free(dt->depth_interval_index);
-    }
-    dt->depth_interval_index = malloc(sizeof(size_t) * (dt->n_ranges + 1));
-
-    size_t i_range_old = dt->t[0].d.i;
-    dt->depth_interval_index[i_range_old] = 0;
-    if(dt->depth_increases) {
-        for(size_t i = 1; i < dt->n; i++) {
-            const des *des = &(dt->t[i]);
-            if(des->d.i > i_range_old) { /* index increases (des table has increasing depth) */
-                for(size_t i_range = i_range_old + 1; i_range <= des->d.i && i_range < dt->n_ranges; i_range++) { /* Handles indices stop_step skips over (no thickness between them) */
-                    dt->depth_interval_index[i_range] = i - 1;
-                }
-                i_range_old = des->d.i;
-            }
-        }
-        dt->depth_interval_index[dt->n_ranges] = dt->n - 1; /* Last point, last index */
-    } else {
-        for(size_t i = 1; i < dt->n; i++) {
-            const des *des = &(dt->t[i]);
-            if(des->d.i < i_range_old) {
-                for(size_t i_range = des->d.i; i_range < i_range_old; i_range++) {
-                    dt->depth_interval_index[i_range + 1] = i;
-                }
-                i_range_old = des->d.i;
-            }
-        }
-        dt->depth_interval_index[i_range_old] = dt->n - 1;
-    }
-}
-
-void des_table_print(FILE *f, const des_table *dt) {
-    if(!dt) {
-        return;
-    }
-    fprintf(f, "DES    i d.i        d.x        d.E      d.S\n");
-    for(size_t i = 0; i < dt->n; i++) {
-        const des *des = &(dt->t[i]);
-        fprintf(f, "DES %4zu %3zu %10.3lf %10.3lf %8.3lf\n", i, des->d.i, des->d.x / C_TFU, des->E / C_KEV, sqrt(des->S) / C_KEV);
-    }
-    fprintf(f, "DES DEPTH INCREASES: %s\n", dt->depth_increases?"TRUE":"FALSE");
-    if(dt->depth_interval_index) {
-        for(size_t i = 0; i < dt->n_ranges; i++) {
-            fprintf(f, "DES INDEX %3zu [%4zu, %4zu]\n", i, dt->depth_interval_index[i], dt->depth_interval_index[i + 1]);
-        }
-    } else {
-        fprintf(f, "DES INDEX DOES NOT EXIST\n");
-    }
-}
-
-depth des_table_find_depth(const des_table *dt, size_t *i_des, depth depth_prev, ion *incident) {
-    size_t i;
-    assert(dt);
-    assert(dt->n > 0);
-    double E = incident->E;
-#ifdef JABS_DEBUG_DES
-    fprintf(stderr, "Where is E = %g keV in DES table? Start index %zu. Previous depth %g tfu (range %zu)\n",
-            E / C_KEV, *i_des, depth_prev.x / C_TFU, depth_prev.i);
-#endif
-    for(i = *i_des; i < dt->n - 1; i++) {
-        if(dt->t[i].d.i != dt->t[i + 1].d.i) { /* This means last point of this layer */
-            E = dt->t[i].E;
-            *i_des = i + 1; /* The +1 prevents stopping at this same layer boundary on the next call */
-#ifdef JABS_DEBUG_DES
-            fprintf(stderr, "Layer boundary at %g tfu. Setting E = %g keV and i_des = %zu. index = %zu\n", dt->t[i].d.x / C_TFU, E / C_KEV, *i_des, dt->depth_interval_index[dt->t[i + 1].d.i]);
-#endif
-            break;
-        }
-        if(dt->t[i].E < E) {/* i is the index of the first element in dt->t that has energy below E. So i-1 should have energy above E. */
-#ifdef JABS_DEBUG_DES
-            fprintf(stderr, "Breaking, i_des = i = %zu, because first element %.12e J (%g keV) <  E = %.12e J (%g keV)\n", i, dt->t[i].E, dt->t[i].E / C_KEV,  E, E / C_KEV);
-#endif
-            *i_des = i;
-            break;
-        }
-    }
-    if(i == dt->n - 1) {
-        if(E < dt->t[i].E) {
-#ifdef JABS_DEBUG_DES
-            fprintf(stderr, "Energy %g keV is below last point in table (%g keV). Changing energy.\n", E / C_KEV, dt->t[i].E / C_KEV);
-#endif
-            E = dt->t[i].E;
-        }
-    } else if(i == 0) {
-        if(E > dt->t[i].E) {
-#ifdef JABS_DEBUG_DES
-            fprintf(stderr, "Energy %g keV is above first point in table (%g keV). Changing energy.\n", E / C_KEV, dt->t[i].E / C_KEV);
-#endif
-            E = dt->t[i].E;
-        }
-        i = 1;
-    }
-    assert(i > 0);
-    const des *des_low = &dt->t[i - 1]; /* Closer to surface, higher energy */
-    const des *des_high = &dt->t[i]; /* Deeper, lower energy */
-#ifdef JABS_DEBUG_DES
-    fprintf(stderr, "dt->t[%zu] = %g tfu (i = %zu), dt->t[%zu] = %g tfu (i = %zu), depth_prev = %g tfu (i = %zu)\n",
-            i - 1,  des_low->d.x / C_TFU, des_low->d.i, i, des_high->d.x / C_TFU, des_high->d.i, depth_prev.x / C_TFU, depth_prev.i);
-#endif
-
-    double E_diff = E - des_low->E;
-    double E_interval = des_high->E - des_low->E;
-    double S_interval = des_high->S - des_low->S;
-    double d_interval = depth_diff(des_low->d, des_high->d);
-    double frac;
-    if(fabs(E_interval) < 0.1 * C_EV) { /* Prevent div by zero */
-        frac = 0.0;
-    } else {
-        frac = (E_diff / E_interval); /* zero if close to low bin */
-    }
-    assert(frac >= 0.0 && frac <= 1.0);
-    depth d_out;
-    if(depth_prev.i != des_high->d.i) { /* First point after crossing layer */
-        d_out.i = des_high->d.i;
-    } else {
-        d_out.i = depth_prev.i;
-    }
-    d_out.x = des_low->d.x + frac * (d_interval); /* Linear interpolation */
-#if 0
-    incident->E = des_low->E + frac * E_interval;
-#else
-    incident->E = E;
-#endif
-    incident->S = des_low->S + frac * S_interval;
-#ifdef JABS_DEBUG_DES
-    fprintf(stderr, "d_out = %g tfu (i = %zu), E = %g keV, S = %g keV\n", d_out.x / C_TFU, d_out.i, incident->E / C_KEV, sqrt(incident->S) * C_FWHM / C_KEV);
-#endif
-    return d_out;
-}
-
-des_table *des_table_compute(const jabs_stop *stop, const jabs_stop *stragg, const sim_calc_params *scp, const sample *sample, const ion *incident, depth depth_start, double emin) {
-    ion ion = *incident;
-    des_table *dt = des_table_init(DES_TABLE_INITIAL_ALLOC);
-    size_t i = 0;
-    depth d_before;
-    depth d_after = depth_start;
-    dt->depth_increases = (ion.inverse_cosine_theta > 0.0);
-    assert(ion.inverse_cosine_theta <= -1.0 || ion.inverse_cosine_theta >= 1.0);
-#ifdef DEBUG
-    fprintf(stderr, "Computing DES table. start_depth = %g tfu, E = %g keV, angle in sample %g deg (1/cos = %.6lf).\n", depth_start.x / C_TFU, ion.E / C_KEV,
-            ion.theta / C_DEG, ion.inverse_cosine_theta);
-#endif
-    do {
-        if(i == dt->n) {
-#ifdef DEBUG
-            fprintf(stderr, "DES table reallocation, size %zu reached.\n", dt->n);
-#endif
-            if(des_table_realloc(dt, dt->n * 2)) {
-                break;
-            }
-        }
-        d_before = d_after;
-        des *des = &dt->t[i];
-        des->E = ion.E;
-        des->S = ion.S;
-        des->d = d_before;
-        if((ion.inverse_cosine_theta > 0.0 && d_before.x >= sample->thickness - DEPTH_TOLERANCE) || (ion.inverse_cosine_theta < 0.0 && d_before.x < DEPTH_TOLERANCE)) {
-#ifdef DEBUG
-            fprintf(stderr, "DES table calculation stops at %g tfu (i = %zu).\n", d_before.x / C_TFU, i);
-#endif
-            i++; /* TODO: Safe? */
-            break;
-        }
-        d_after = stop_step(stop, stragg, &ion, sample, d_before, stop_step_calc_incident(scp, &ion));
-        i++;
-    } while(ion.E > emin);
-    if(dt->n) {
-        des_table_realloc(dt, i); /* Shrinks to size */
-        if(i > 0 ) {
-            dt->n_ranges = GSL_MAX(depth_start.i, dt->t[i - 1].d.i) + 1;
-        } else {
-            dt->n_ranges = 0;
-        }
-        des_table_rebuild_index(dt);
-        return dt;
-    } else {
-        des_table_free(dt);
-        return NULL;
-    }
-}
-
-void des_set_ion(const des *des, ion *ion) {
-    ion->E = des->E;
-    ion->S = des->S;
-}
-
 double cross_section_straggling_fixed(const sim_reaction *sim_r, const prob_dist *pd, double E, double S) {
     const double std_dev = sqrt(S);
     double cs_sum = 0.0;
@@ -521,7 +268,7 @@ void exit_from_sample(ion *p, const depth depth_start, const sim_workspace *ws, 
         if(p->inverse_cosine_theta < 0.0 && d.x <= DEPTH_TOLERANCE) { /* Exit (surface, front of sample) */
             break;
         }
-        double E_step = stop_step_calc_exiting(ws->params, p);
+        double E_step = stop_step_calc(&ws->params->exiting_stop_params, p);
         depth d_after = stop_step(&ws->stop, &ws->stragg, p, sample, d, E_step);
         if(p->E < ws->emin) {
 #ifdef DEBUG_REACTION
@@ -533,26 +280,6 @@ void exit_from_sample(ion *p, const depth depth_start, const sim_workspace *ws, 
         }
         d = d_after;
     }
-}
-
-double stop_step_calc_incident(const sim_calc_params *params, const ion *ion) {
-    if(params->incident_stop_step > 0.0) {
-        return params->incident_stop_step;
-    }
-    double step = params->incident_stop_step_sigmas * sqrt(ion->S);
-    step = GSL_MIN_DBL(step, params->incident_stop_step_max);
-    step = GSL_MAX_DBL(step, params->incident_stop_step_min);
-    return step;
-}
-
-double stop_step_calc_exiting(const sim_calc_params *params, const ion *ion) {
-    if(params->exiting_stop_step > 0.0) {
-        return params->exiting_stop_step;
-    }
-    double step = params->exiting_stop_step_sigmas * sqrt(ion->S);
-    step = GSL_MIN_DBL(step, params->exiting_stop_step_max);
-    step = GSL_MAX_DBL(step, params->exiting_stop_step_min);
-    return step;
 }
 
 void simulate_reaction(const ion *incident, const depth depth_start, sim_workspace *ws, const sample *sample, const des_table *dt, const geostragg_vars *g, sim_reaction *sim_r) {
@@ -588,7 +315,7 @@ void simulate_reaction(const ion *incident, const depth depth_start, sim_workspa
             last = TRUE;
         }
         d_after = des_table_find_depth(dt, &i_des, d_before, &ion1); /* Does this handle E below min? */
-        if(i_brick == 0 || d_after.i > d_before.i) { /* There was a layer (depth range) crossing. If stop_step() took this into account when making DES table the only issue is the .i index. depth (.x) is not changed. */
+        if(i_brick == 0 || d_after.i > d_before.i) { /* There was a layer (depth range) crossing. If step() took this into account when making DES table the only issue is the .i index. depth (.x) is not changed. */
 #ifndef NO_SKIP_EMPTY_RANGES
             double conc_start = *sample_conc_bin(sample, d_after.i, sim_r->i_isotope);
             double conc_stop = *sample_conc_bin(sample, d_after.i + 1, sim_r->i_isotope);
@@ -1178,7 +905,7 @@ int simulate_with_ds(sim_workspace *ws) {
         double E_front = ion1.E;
         if(E_front <= ws->emin)
             break;
-        double E_step = ws->params->ds_incident_stop_step_factor * stop_step_calc_incident(ws->params, &ion1);
+        double E_step = ws->params->ds_incident_stop_step_factor * stop_step_calc(&ws->params->incident_stop_params, &ion1);
         if(E_front - E_step <= ws->emin) { /* Fix last step to be "just enough" */
             last = TRUE;
             E_step = E_front - ws->emin;
