+++ src/jabs.h
@@ -22,34 +22,9 @@
 #include "reaction.h"
 #include "geostragg.h"
 #include "stop.h"
+#include "des.h"
 
-typedef struct {
-    depth d; /* Depth */
-    double E; /* Energy */
-    double S; /* Straggling */
-} des; /* DES = Depth, Energy, Straggling */
-
-typedef struct {
-    des *t; /* array, n elements. E decreasing, d either increases (ion going deeper) or decreases. S can do whatever S does. */
-    int depth_increases; /* TRUE if d increases (going deeper) */
-    size_t n;
-    size_t n_ranges;
-    size_t *depth_interval_index; /* table, size same as number of sample ranges (as given in des_table_compute()). Array stores location i of t[i].d.x == sample->range[i_range].x  */
-} des_table; /* Depth, energy, straggling table */
-
-int des_table_realloc(des_table *dt, size_t n);
-void des_table_free(des_table *dt);
-des_table *des_table_compute(const jabs_stop *stop, const jabs_stop *stragg, const sim_calc_params *scp, const sample *sample, const ion *incident, depth depth_start, double emin);
-size_t des_table_size(const des_table *dt);
-des *des_table_element(const des_table *dt, size_t i);
-void des_table_rebuild_index(des_table *dt); /* called by des_table_compute() after setting values to table and before any other function can be used */
-void des_table_print(FILE *f, const des_table *dt);
-depth des_table_find_depth(const des_table *dt, size_t *i_des, depth depth_prev, ion *incident); /* Returns depth at given incident->E, or the next layer boundary, starting search from i_des in DES table. Updates i_des, incident->E and ->S. */
-inline const des *des_table_min_energy_bin(const des_table *dt) {return &(dt->t[dt->n - 1]);}
-void des_set_ion(const des *des, ion *ion);
 void exit_from_sample(ion *p, depth depth_start, const sim_workspace *ws, const sample *sample);
-double stop_step_calc_incident(const sim_calc_params *params, const ion *ion);
-double stop_step_calc_exiting(const sim_calc_params *params, const ion *ion);
 int simulate(const ion *incident, depth depth_start, sim_workspace *ws, const sample *sample);
 void simulate_reaction(const ion *incident, depth depth_start, sim_workspace *ws, const sample *sample, const des_table *dt, const geostragg_vars *g, sim_reaction *sim_r);
 void simulate_init_reaction(sim_reaction *sim_r, const sample *sample, const geostragg_vars *g, double E_min, double E_max);
