+++ src/script_command.c
@@ -1230,34 +1230,34 @@ script_command *script_commands_create(struct script_session *s) {
             {JIBAL_CONFIG_VAR_BOOL,   "erd",                         &sim->erd,                                 NULL},
             {JIBAL_CONFIG_VAR_BOOL,   "rbs",                         &sim->rbs,                                 NULL},
             {JIBAL_CONFIG_VAR_SIZE,   "maxiter",                     &fit->n_iters_max,                         NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "xtolerance",                  &fit->xtol,                                NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "gtolerance",                  &fit->gtol,                                NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "ftolerance",                  &fit->ftol,                                NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "chisq_tolerance",             &fit->chisq_tol,                               NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "chisq_fast_tolerance",        &fit->chisq_fast_tol,                          NULL},
-            {JIBAL_CONFIG_VAR_SIZE,   "depthsteps_max",              &sim->params->depthsteps_max,                  NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "ds",                          &sim->params->ds,                              NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "rk4",                         &sim->params->rk4,                             NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "brick_width_sigmas",          &sim->params->brick_width_sigmas,              NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "sigmas_cutoff",               &sim->params->sigmas_cutoff,                   NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step",          &sim->params->incident_stop_step,              NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_sigmas",   &sim->params->incident_stop_step_sigmas,       NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_min",      &sim->params->incident_stop_step_min,          NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_max",      &sim->params->incident_stop_step_max,          NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step",           &sim->params->exiting_stop_step,               NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_sigmas",    &sim->params->exiting_stop_step_sigmas,        NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_min",       &sim->params->exiting_stop_step_min,           NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_max",       &sim->params->exiting_stop_step_max,           NULL},
-            {JIBAL_CONFIG_VAR_UNIT,   "ds_incident_stop_step_factor",&sim->params->ds_incident_stop_step_factor,    NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "nuclear_stopping_accurate",   &sim->params->nuclear_stopping_accurate,       NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "mean_conc_and_energy",        &sim->params->mean_conc_and_energy,            NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "geostragg",                   &sim->params->geostragg,                       NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "beta_manual",                 &sim->params->beta_manual,                     NULL},
-            {JIBAL_CONFIG_VAR_SIZE,   "cs_n_stragg_steps",           &sim->params->cs_n_stragg_steps,               NULL},
-            {JIBAL_CONFIG_VAR_BOOL,   "gaussian_accurate",           &sim->params->gaussian_accurate,               NULL},
-            {JIBAL_CONFIG_VAR_SIZE,   "int_cs_max_intervals",        &sim->params->int_cs_max_intervals,            NULL},
-            {JIBAL_CONFIG_VAR_DOUBLE, "int_cs_accuracy",             &sim->params->int_cs_accuracy,                 NULL},
-            {JIBAL_CONFIG_VAR_SIZE,   "int_cs_stragg_max_intervals", &sim->params->int_cs_stragg_max_intervals,     NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "xtolerance",                  &fit->xtol,                                 NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "gtolerance",                  &fit->gtol,                                 NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "ftolerance",                  &fit->ftol,                                 NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "chisq_tolerance",             &fit->chisq_tol,                            NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "chisq_fast_tolerance",        &fit->chisq_fast_tol,                       NULL},
+            {JIBAL_CONFIG_VAR_SIZE,   "depthsteps_max",              &sim->params->depthsteps_max,               NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "ds",                          &sim->params->ds,                           NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "rk4",                         &sim->params->rk4,                          NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "brick_width_sigmas",          &sim->params->brick_width_sigmas,           NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "sigmas_cutoff",               &sim->params->sigmas_cutoff,                NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step",          &sim->params->incident_stop_params.step,    NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_sigmas",   &sim->params->incident_stop_params.sigmas,  NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_min",      &sim->params->incident_stop_params.min,     NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "incident_stop_step_max",      &sim->params->incident_stop_params.max,     NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step",           &sim->params->exiting_stop_params.step,     NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_sigmas",    &sim->params->exiting_stop_params.sigmas,   NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_min",       &sim->params->exiting_stop_params.min,      NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "exiting_stop_step_max",       &sim->params->exiting_stop_params.max,      NULL},
+            {JIBAL_CONFIG_VAR_UNIT,   "ds_incident_stop_step_factor",&sim->params->ds_incident_stop_step_factor, NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "nuclear_stopping_accurate",   &sim->params->nuclear_stopping_accurate,    NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "mean_conc_and_energy",        &sim->params->mean_conc_and_energy,         NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "geostragg",                   &sim->params->geostragg,                    NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "beta_manual",                 &sim->params->beta_manual,                  NULL},
+            {JIBAL_CONFIG_VAR_SIZE,   "cs_n_stragg_steps",           &sim->params->cs_n_stragg_steps,            NULL},
+            {JIBAL_CONFIG_VAR_BOOL,   "gaussian_accurate",           &sim->params->gaussian_accurate,            NULL},
+            {JIBAL_CONFIG_VAR_SIZE,   "int_cs_max_intervals",        &sim->params->int_cs_max_intervals,         NULL},
+            {JIBAL_CONFIG_VAR_DOUBLE, "int_cs_accuracy",             &sim->params->int_cs_accuracy,              NULL},
+            {JIBAL_CONFIG_VAR_SIZE,   "int_cs_stragg_max_intervals", &sim->params->int_cs_stragg_max_intervals,  NULL},
             {JIBAL_CONFIG_VAR_DOUBLE, "int_cs_stragg_accuracy",      &sim->params->int_cs_stragg_accuracy,      NULL},
             {JIBAL_CONFIG_VAR_BOOL,   "cs_adaptive",                 &sim->params->cs_adaptive,                 NULL},
             {JIBAL_CONFIG_VAR_UNIT,   "cs_energy_step_max",          &sim->params->cs_energy_step_max,          NULL},
